Move components from index to separate components
